+++ sdk/src/androidTest/java/ly/count/android/sdk/UtilsTests.java
@@ -10,6 +10,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import static org.mockito.Mockito.mock;
 
 @RunWith(AndroidJUnit4.class)
 public class UtilsTests {
@@ -142,5 +143,61 @@ public class UtilsTests {
         Assert.assertEquals(123, segm.get("2"));
         Assert.assertEquals(345.33d, segm.get("3"));
         Assert.assertEquals(false, segm.get("4"));
+    }
+
+    /**
+     * Make sure that nothing bad happens when providing null segmentation
+     */
+    @Test
+    public void testTruncateSegmentationValues_null() {
+        Utils.truncateSegmentationValues(null, 10, "someTag", mock(ModuleLog.class));
+        Assert.assertTrue(true);
+    }
+
+    /**
+     * Make sure that nothing bad happens when providing empty segmentation
+     */
+    @Test
+    public void testTruncateSegmentationValues_empty() {
+        Map<String, Object> values = new HashMap<>();
+        Utils.truncateSegmentationValues(values, 10, "someTag", mock(ModuleLog.class));
+        Assert.assertTrue(true);
+    }
+
+    /**
+     * Make sure that nothing bad happens when providing segmentation with values under limit
+     */
+    @Test
+    public void testTruncateSegmentationValues_underLimit() {
+        Map<String, Object> values = new HashMap<>();
+        values.put("a1", "1");
+        values.put("a2", "2");
+        values.put("a3", "3");
+        values.put("a4", "4");
+        Utils.truncateSegmentationValues(values, 6, "someTag", mock(ModuleLog.class));
+
+        Assert.assertEquals(4, values.size());
+        Assert.assertEquals("1", values.get("a1"));
+        Assert.assertEquals("2", values.get("a2"));
+        Assert.assertEquals("3", values.get("a3"));
+        Assert.assertEquals("4", values.get("a4"));
+    }
+
+    /**
+     * Make sure that values are truncated when they are more then the limit
+     */
+    @Test
+    public void testTruncateSegmentationValues_aboveLimit() {
+        Map<String, Object> values = new HashMap<>();
+        values.put("a1", "1");
+        values.put("a2", "2");
+        values.put("a3", "3");
+        values.put("a4", "4");
+        Utils.truncateSegmentationValues(values, 2, "someTag", mock(ModuleLog.class));
+
+        Assert.assertEquals(2, values.size());
+        //after inspecting what is returned in the debugger, it should have the values of "a2" and "a4"
+        Assert.assertEquals("2", values.get("a2"));
+        Assert.assertEquals("4", values.get("a4"));
     }
 }