+++ CHANGELOG.md
@@ -6,6 +6,13 @@
   * ! Minor breaking change ! Automatic "AdvertisingId" attribution has been removed from the SDK. A call is added for recording provided attribution values.
   * ! Minor breaking change ! Automatic install referrer attribution in play stores has been removed. A call is added for recording such provided values.
   * ! Minor breaking change ! Orientation tracking is now enabled by default. Can be turned off during init.
+  * ! Minor breaking change ! Internal SDK limits are now introduced. These limits are adjustable at initialization:
+    - Maximum length of all keys is now 128 characters by default.
+    - Maximum length of all values is now 256 characters by default.
+    - Maximum amount of provided segmentation in now 30 entries by default.
+    - Maximum amount of breadcrumbs that can be recorded at once is now 100 by default.
+    - Maximum stack trace lines per thread is now 30 by default.
+    - Maximum stack trace line length is now 200 by default.
 
   * Device ID can now be changed when no consent is given
   * Push notification now display/use the sent badge number. It's visualization depends on the launcher.